+++ package.js
@@ -2,18 +2,18 @@ Package.describe({
   summary: "Add profile.online field and real-time updates to Meteor.users."
 });
 
-var both = ['client', 'server'];
-
 Package.on_use( function(api) {
     api.use('accounts-base', ['client','server']);
     api.use('coffeescript', 'server');
     api.add_files('user_status.coffee', 'server');
+
+    api.export(['UserSessions', 'UserStatus'], 'server');
 });
 
 Package.on_test( function(api) {
-    api.use('user-status', both);
-    api.use('test-helpers', both);
-    api.use('tinytest', both);
+    api.use('user-status');
+    api.use('test-helpers');
+    api.use('tinytest');
 
     // TODO add test cases
 });